Split out non-localizable version data from versionInfo into a new buildVersion module. This allows importing of buildVersion in places where the version is needed before languageHandler is initialized. Specifically, this is now used in logHandler to stop versionInfo being accidentally imported too early.
